requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
mysql-connector-python
python-dotenv
sqlalchemy